requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
bitarray==2.9.2
blinker==1.6.2
cachelib==0.11.0
celery==5.3.6
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
convertapi==1.6.0
cytoolz==0.12.3
dnspython==2.5.0
easygoogletranslate==0.0.4
eth-account==0.11.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
exceptiongroup==1.2.0
Flask==2.2.5
Flask-Cors==3.0.10
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
frozenlist==1.3.3
g4f==0.2.1.3
grpcio==1.53.0
grpcio-tools==1.53.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
lru-dict==1.2.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.26.4
oauthlib==2.1.0
openai==0.27.6
opencv-python==4.7.0.72
outcome==1.3.0.post0
packaging==23.2
param==1.13.0
parsimonious==0.9.0
Pillow==9.5.0
prompt-toolkit==3.0.43
protobuf==4.21.12
pycryptodome==3.20.0
pydispo==20.10b1
pymongo==4.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
redis==5.0.2
referencing==0.33.0
regex==2023.12.25
requests==2.30.0
requests-oauthlib==1.1.0
rlp==4.0.0
rpds-py==0.17.1
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toolz==0.12.1
tqdm==4.65.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.0.2
vine==5.1.0
waitress==2.1.2
wcwidth==0.2.13
web3==6.15.1
websockets==12.0
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.2
zipp==3.15.0
